django/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
chardet==2.2.1
colorama==0.2.5
Django==1.8.3
djangorestframework==3.2.4
elasticsearch==1.6.0
gunicorn==19.4.1
html5lib==0.999
kombu==3.0.26
psycopg2==2.6.1
pymssql==2.1.3
PyMySQL==0.7.2
python-dateutil==2.4.2
pytz==2015.4
redis==2.10.3
requests==2.2.1
six==1.5.2
urllib3==1.10.4
wheel==0.24.0
xmltodict==0.9.2
bugsnag==2.5.0
